chore: renamed class and imports from into their dependents
